feat(eval): add evaluation helpers for codec testing
Added lightweight helper functions for simple quality evaluation:

- `evaluate_single()`: Evaluate a single image against a reference
- `assert_quality()`: Assert quality thresholds (for CI tests)
- `assert_perception_level()`: Assert semantic quality levels

These helpers make it easy for zen* projects to:
- Quickly check encoded image quality during development
- Add quality regression tests to CI
- Compare images without setting up full EvalSession

Changes:
- Added eval/helpers.rs with conversion utilities
- Added QualityBelowThreshold error variant
- Fixed DimensionMismatch error to use usize instead of u32
- Added tests for all helper functions

Example usage:
```rust
use codec_eval::eval::helpers::*;
use codec_eval::metrics::MetricConfig;

// Evaluate quality
let config = MetricConfig::perceptual();
let result = evaluate_single(&reference, &encoded, &config)?;

// Assert in tests
assert_quality(&reference, &encoded, Some(80.0), Some(0.002))?;

// Semantic quality levels
assert_perception_level(&reference, &encoded, PerceptionLevel::Subtle)?;
```

+++ src/metrics/butteraugli.rs
@@ -43,8 +43,8 @@ pub fn calculate_butteraugli(
 ) -> Result<f64> {
     if reference.len() != test.len() {
         return Err(Error::DimensionMismatch {
-            expected: (width as u32, height as u32),
-            actual: ((test.len() / 3 / height) as u32, height as u32),
+            expected: (width, height),
+            actual: (test.len() / 3 / height, height),
         });
     }
 
@@ -96,8 +96,8 @@ pub fn calculate_butteraugli_with_intensity(
 ) -> Result<f64> {
     if reference.len() != test.len() {
         return Err(Error::DimensionMismatch {
-            expected: (width as u32, height as u32),
-            actual: ((test.len() / 3 / height) as u32, height as u32),
+            expected: (width, height),
+            actual: (test.len() / 3 / height, height),
         });
     }
